Track punch office/location independent of IP address
A new option '$audit_office' will enable setting a long-lived browser
cookie in the office admin screen which identifies the office of that
particular browser. This name will display in the top navigation bar and
will be stored in the punch log for any users who punch in on that browser.

The primary use is as an automatic and convenient method of tracking who is
in the building when you have multiple locations and some remote employees.
IP address currently serves this role badly because numeric addresses are
not friendly and are not reliable due to dynamic addresses and VPNs.

This feature also provides mild protection against remote punching since
setting the browser office (through the application) requires admin access.
Advanced users could probably figure out how to bypass this.